+++ CHANGELOG.md
@@ -47,6 +47,10 @@
 
 * Add new `allowed_numbers` option to the `no_magic_numbers` rule.  
   [Martin Redington](https://github.com/mildm8nnered)
+* Add new `unnest_switches_using_tuple` opt-in rule that triggers when
+  nested switches are encountered that reference the same variable. These
+  can be replaced by a switch on a tuple.  
+  [Alfons Hoogervorst](https://github.com/snofla/SwiftLint)
 
 ### Bug Fixes